refactor: cleanup playwright test
