+++ tests/components/color-picker.test.tsx
